Bugfix/alex/small issues

+++ src/database/migrations/V8__create_order_details_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "OrderDetails" (
-  "OrderDetailID" bigint PRIMARY KEY NOT NULL,
+  "OrderDetailID" SERIAL PRIMARY KEY NOT NULL,
   "OrderID" bigint NOT NULL,
   "ProductID" bigint NOT NULL,
   "Quantity" int NOT NULL,
