+++ src/database/migrations/V9__create_payments_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "Payments" (
-  "PaymentID" bigint PRIMARY KEY NOT NULL,
+  "PaymentID" SERIAL PRIMARY KEY NOT NULL,
   "OrderID" bigint NOT NULL,
   "Amount" decimal(8,2) NOT NULL,
   "PaymentDate" date NOT NULL,
